Added nodes for the AST and fixed some edge cases
The core idea of the project is that we will turn the program into a big
AST which will behave as the symbol table and then use it for analysis
and some optimizations, with this we can reduce unnecessary nodes where
the remaining will be compiled to bytecode. We can track variable usage
and eliminate dead code which amazingly will be just removing nodes from
the AST. I've not thought this very well but if my clculations are
correct this will turn out fine.