Added some errors and fixed yesterdays bug
The bug was caused by the lack of the GROUP precedence, that's all.
As promised, I've added errors for the expression parsing. I started the
statement parsing step and realized my design is pretty poor, that is
the errors and I have not settled in the syntax of the language yet, for
the expressions, they now work perfectly and as a test I added new
operators, +=, -=, *= and /= and they process was so simple,No new nodes
though since they are just syntactic sugars with some extra constraints.

I need to sit on some work; define the syntax, understand how to attach
tokens to the various statements and tokens to help in debugging in the
semantic analysis phase and others. For example, why Im I supporting c
style for loops and yet I was planning to add iterators and generators
to this language, I think the for loop should be like the one in python;
example `for(value : iterable)` where value is a local variable to the
for loop and iterable conforms to some defined iterator interface. This
is why I need to plan this out. I will delete a lot of code and
reimplement them with this structure in mind. For now, I'll take a
break, a design break, put my shit together and come back with a game
plan, a good one.

My hopes of this project are pretty high and I wouldn't want such
incompetence ruin it all. I want a simple language with modern features
like iterators, generators, asynchronous, error handling, compiled and
optimized. For the optimization part, we'll be deading with the visitor
pattern as originally planned and all other phases including the
compiler as illustrated in the YAP/exprlang project.